requirements.txt
# FastAPI e servidor ASGI
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Autenticação e segurança
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Validação de emails
email-validator==2.1.0

# CORS
python-dotenv==1.0.0

# Logging e utilitários
pydantic==2.5.0
pydantic-settings==2.1.0
